+++ test/unit/config-manager-test.js
@@ -126,6 +126,16 @@ describe('config-manager', function() {
       })
     })
 
+    describe('#updateTx', function() {
+      it('replaces the tx with the same id', function() {
+        configManager.addTx({ id: '1', status: 'unconfirmed' })
+        configManager.addTx({ id: '2', status: 'confirmed' })
+        configManager.updateTx({ id: '1', status: 'blah', hash: 'foo' })
+        var result = configManager.getTx('1')
+        assert.equal(result.hash, 'foo')
+      })
+    })
+
     describe('#unconfirmedTxs', function() {
       it('returns unconfirmed txs in a hash', function() {
         configManager.addTx({ id: '1', status: 'unconfirmed' })